factoring of some ZAM header files for better modularity
