+++ src/script_opt/ZAM/BuiltInSupport.h
@@ -1,5 +1,7 @@
 // See the file "COPYING" in the main distribution directory for copyright.
 
+#pragma once
+
 #include "zeek/Desc.h"
 #include "zeek/Expr.h"
