+++ src/script_opt/ZAM/IterInfo.h
@@ -4,10 +4,9 @@
 
 #pragma once
 
-#include "zeek/Dict.h"
 #include "zeek/Val.h"
 #include "zeek/ZeekString.h"
-#include "zeek/script_opt/ZAM/ZInst.h"
+#include "zeek/script_opt/ZAM/ZInstAux.h"
 
 namespace zeek::detail {
